app/requirements.txt
Flask==2.1.0
PyMySQL==1.0.2
urllib3
mysql-connector